Allow build changes to be deployed
The deploy task will, if there are changes to the compiled assets,
attempt to commit and push those. It just so happens that both a) Circle
CI didn't have git configured and b) we did not have any changes to
assets since we started deploying from Circle.

This has been an outstanding issue for a while; this should fix
deployment in those cases.